changes regarding listing orders with payment status paid on online payment in App and web

+++ app/Http/Controllers/Api/v1/OrderController.php
@@ -312,12 +312,17 @@ class OrderController extends BaseController {
                     $order->scheduled_date_time = $cart->schedule_type == 'schedule' ? $cart->scheduled_date_time : null;
                     $order->subscription_discount = $total_subscription_discount;
                     $order->payable_amount = $payable_amount;
+                    if ( ($payable_amount == 0) || (($request->has('transaction_id')) && (!empty($request->transaction_id))) ) {
+                        $order->payment_status = 1;
+                    }
                     $order->save();
                     $this->sendSuccessSMS($request, $order);
-                    Cart::where('id', $cart->id)->update(['schedule_type' => NULL, 'scheduled_date_time' => NULL]);
-                    CartCoupon::where('cart_id', $cart->id)->delete();
-                    CartProduct::where('cart_id', $cart->id)->delete();
-                    CartProductPrescription::where('cart_id', $cart->id)->delete();
+                    if($request->payment_option_id != 7){ // if not mobbex
+                        Cart::where('id', $cart->id)->update(['schedule_type' => NULL, 'scheduled_date_time' => NULL]);
+                        CartCoupon::where('cart_id', $cart->id)->delete();
+                        CartProduct::where('cart_id', $cart->id)->delete();
+                        CartProductPrescription::where('cart_id', $cart->id)->delete();
+                    }
                     if (($request->payment_option_id != 1) && ($request->payment_option_id != 2) && ($request->has('transaction_id')) && (!empty($request->transaction_id))) {
                         Payment::insert([
                             'date' => date('Y-m-d'),
@@ -326,29 +331,31 @@ class OrderController extends BaseController {
                             'balance_transaction' => $order->payable_amount,
                         ]);
                     }
-                    $code = $request->header('code');
                     $order = $order->with(['vendors:id,order_id,vendor_id', 'user_vendor'])->where('order_number', $order->order_number)->first();
-                    if (!empty($order->vendors)) {
-                        foreach ($order->vendors as $vendor_value) {
-                            $vendor_order_detail = $this->minimize_orderDetails_for_notification($order->id, $vendor_value->vendor_id);
-                            $user_vendors = UserVendor::where(['vendor_id' => $vendor_value->vendor_id])->pluck('user_id');
-                            $this->sendOrderPushNotificationVendors($user_vendors, $vendor_order_detail, $code);
+                    if($request->payment_option_id != 7){ // if not mobbex
+                        $code = $request->header('code');
+                        if (!empty($order->vendors)) {
+                            foreach ($order->vendors as $vendor_value) {
+                                $vendor_order_detail = $this->minimize_orderDetails_for_notification($order->id, $vendor_value->vendor_id);
+                                $user_vendors = UserVendor::where(['vendor_id' => $vendor_value->vendor_id])->pluck('user_id');
+                                $this->sendOrderPushNotificationVendors($user_vendors, $vendor_order_detail, $code);
+                            }
                         }
-                    }
-                    $vendor_order_detail = $this->minimize_orderDetails_for_notification($order->id);
-                    $super_admin = User::where('is_superadmin', 1)->pluck('id');
-                    $this->sendOrderPushNotificationVendors($super_admin, $vendor_order_detail, $code);
-                    // $user_admins = User::where(function ($query) {
-                    //     $query->where(['is_superadmin' => 1]);
-                    // })->pluck('id')->toArray();
-                    // $user_vendors = [];
-                    // if (!empty($order->user_vendor) && count($order->user_vendor) > 0) {
-                    //     $user_vendors = $order->user_vendor->pluck('user_id')->toArray();
-                    // }
-                    // $order->admins = array_unique(array_merge($user_admins, $user_vendors));
+                        $vendor_order_detail = $this->minimize_orderDetails_for_notification($order->id);
+                        $super_admin = User::where('is_superadmin', 1)->pluck('id');
+                        $this->sendOrderPushNotificationVendors($super_admin, $vendor_order_detail, $code);
+                        // $user_admins = User::where(function ($query) {
+                        //     $query->where(['is_superadmin' => 1]);
+                        // })->pluck('id')->toArray();
+                        // $user_vendors = [];
+                        // if (!empty($order->user_vendor) && count($order->user_vendor) > 0) {
+                        //     $user_vendors = $order->user_vendor->pluck('user_id')->toArray();
+                        // }
+                        // $order->admins = array_unique(array_merge($user_admins, $user_vendors));
 
-                    // // $this->sendOrderNotification($user->id);
-                    // $this->sendOrderPushNotificationVendors($order->admins, ['id' => $order->id], $code);
+                        // // $this->sendOrderNotification($user->id);
+                        // $this->sendOrderPushNotificationVendors($order->admins, ['id' => $order->id], $code);
+                    }
                     DB::commit();
 
                     return $this->successResponse($order, __('Order placed successfully.'), 201);
